Split off stream and exampleData
